moved x86_64 into a file with preprocessor ifdef
